+++ frontend/app/components/Navigation/UserModal.tsx
@@ -6,9 +6,9 @@ interface UserModalComponentProps {
     modal_id: string;
     title: string;
     text: string;
-    triggerAccept: (a: any) => void
-    triggerValue: any
-    triggerString: string
+    triggerAccept?: (a: any) => void
+    triggerValue?: any
+    triggerString?: string
 }
 
 const UserModalComponent: React.FC<UserModalComponentProps> = ({ title, modal_id, text, triggerAccept, triggerString, triggerValue }) => {
@@ -16,11 +16,13 @@ const UserModalComponent: React.FC<UserModalComponentProps> = ({ title, modal_id
         <dialog id={modal_id} className="modal">
             <div className="modal-box">
                 <h3 className="font-bold text-lg">{title}</h3>
-                <p>{text}</p>
+                <p className='whitespace-pre-wrap'>{text}</p>
                 <div className="modal-action">
                     <form method="dialog">
-                        <button className="btn text-text-verba bg-primary-verba hover:bg-button-hover-verba" onClick={() => { triggerAccept(triggerValue) }} >{triggerString}</button>
-                        <button className="btn text-text-verba bg-warning-verba hover:bg-button-hover-verba ml-2">Cancel</button>
+                        {triggerAccept && triggerValue && triggerString && (
+                            <button className="btn text-text-verba bg-primary-verba border-none hover:bg-button-hover-verba" onClick={() => { triggerAccept(triggerValue) }} >{triggerString}</button>
+                        )}
+                        <button className="btn text-text-verba bg-warning-verba border-none hover:bg-button-hover-verba ml-2">Cancel</button>
                     </form>
                 </div>
             </div>