+++ goldenverba/components/embedding/MiniLMEmbedder.py
@@ -58,8 +58,8 @@ class MiniLMEmbedder(Embedder):
         for document in tqdm(
             documents, total=len(documents), desc="Vectorizing document chunks"
         ):
-            for chunk in document.chunks:
-                chunk.set_vector(self.vectorize_chunk(chunk.text))
+            for chunk in tqdm(document.chunks, total=len(document.chunks), desc="Vectorizing Chunks"):
+                chunk.set_vector(self.vectorize_chunk(document.name + " : " + chunk.text))
 
         return self.import_data(documents, client, logging)